agregado cors

+++ app/Http/Controllers/UserController.php
@@ -10,10 +10,12 @@ use Illuminate\Support\Facades\Hash;
 class UserController extends Controller
 {
     public function index(Request $req){
+        if($req->user()->rol != 'A') return response()->json(['status'=>'failed'], 401);
         return User::all();
     }
 
-    public function get($user){
+    public function get(Request $req, $user){
+        if($req->user()->rol != 'A') return response()->json(['status'=>'failed'], 401);
         $result = User::find($user);
         //$result = DB::table('users')->where('user', '=', $user)->get();
         if($result)
@@ -23,6 +25,7 @@ class UserController extends Controller
     }
 
     public function create(Request $req){
+        if($req->user()->rol != 'A') return response()->json(['status'=>'failed'], 401);
         $this->validate($req, [
             'user'=>'required', 
             'nombre'=>'required',
@@ -43,6 +46,7 @@ class UserController extends Controller
     }
 
     public function update(Request $req, $user){
+        if($req->user()->rol != 'A') return response()->json(['status'=>'failed'], 401);
         $this->validate($req, [
             'user'=>'filled', 
             'nombre'=>'filled',
@@ -59,8 +63,8 @@ class UserController extends Controller
             return response()->json(['status'=>'failed'], 404);
     }
 
-    public function destroy($user){
-        
+    public function destroy(Request $req, $user){
+        if($req->user()->rol != 'A') return response()->json(['status'=>'failed'], 401);
         $datos = User::find($user);
         if(!$datos) return response()->json(['status'=>'failed'], 404);
         $result = $datos->delete();
